pyproject.toml
[project]
name = "crawlforge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "beautifulsoup4>=4.14.3",
    "black>=26.1.0",
    "faiss-cpu>=1.13.2",
    "httpx>=0.28.1",
    "readability-lxml>=0.8.4.1",
    "redis>=7.3.0",
    "sentence-transformers>=5.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]
